Create alphabetical order for blocks' code
