+++ src/test/java/com/epam/prejap/tetris/block/BlocksShapesTest.java
@@ -4,113 +4,47 @@ import org.testng.annotations.DataProvider;
 import org.testng.annotations.Factory;
 import org.testng.annotations.Test;
 
+import java.util.Iterator;
+
 import static org.testng.Assert.assertEquals;
 
+/**
+ * "Generic" test class, to test shapes.
+ * Define block class and its expected image in {@link BlocksShapesData} class
+ * and it will be tested automatically
+ *
+ * @author Nikita Pochapynskyi
+ */
 @Test(groups = "blockShapes")
 public class BlocksShapesTest {
 
-    private final int rowsExpected;
-    private final int rowsActual;
-    private final int colsExpected;
-    private final int colsActual;
-    private final String testedBlock;
-    private final Object[][] dotRepresentation;
+    private final Block block;
+    private final byte[][] image;
 
-    public void numberOfEveryBlockShapeRowsAsExpected() {
-        assertEquals(rowsActual, rowsExpected, "Number of rows for " + testedBlock +
-                "is incorrect. Expected: " + rowsExpected + ", actual: " + rowsActual);
-    }
-
-    public void numberOfEveryBlockShapeColsAsExpected() {
-        assertEquals(colsActual, colsExpected, "Number of cols for " + testedBlock +
-                "is incorrect. Expected: " + colsExpected + ", actual: " + colsActual);
-    }
-
-    @Test(dataProvider = "dotsRepresentation")
-    public void dotAtReturnCorrectValue(int actual, int expected, String message) {
-        String msg = testedBlock + message;
-        assertEquals(actual, expected, msg);
+    @Factory(dataProviderClass = BlocksShapesData.class, dataProvider = "blocks")
+    public BlocksShapesTest(Block block, byte[][] image) {
+        this.block = block;
+        this.image = image;
     }
 
     @DataProvider
-    public Object[][] dotsRepresentation() {
-        return dotRepresentation;
+    public Iterator<Object[]> dotAtData() {
+        return BlocksShapesData.getDotAtDataFor(block.getClass());
     }
 
-    @Factory(dataProvider = "blocks")
-    public BlocksShapesTest(String testedBlock, int rowsActual, int rowsExpected, int colsActual, int colsExpected, Object[][] dotRepresentation) {
-        this.rowsActual = rowsActual;
-        this.rowsExpected = rowsExpected;
-        this.colsActual = colsActual;
-        this.colsExpected = colsExpected;
-        this.testedBlock = testedBlock;
-        this.dotRepresentation = dotRepresentation;
+    public void shouldBeRightQtOfRows() {
+        assertEquals(block.rows(), image.length,
+                String.format("%s: Nr of rows for is incorrect.", block.getClass().getSimpleName()));
     }
 
-    @DataProvider
-    public static Object[][] blocks() {
-        OBlock oBlock = new OBlock();
-        int actualRowsForOBlock = oBlock.rows();
-        int expectedRowsForOBlock = 2;
-        int actualColsForOBlock = oBlock.cols();
-        int expectedColsForOBlock = 2;
-        var oBlockDotRepresentation = new Object[][]{
-                {oBlock.dotAt(0, 0), 1, "incorrect dotAt(0, 0)"},
-                {oBlock.dotAt(0, 1), 1, "incorrect dotAt(0, 1)"},
-                {oBlock.dotAt(1, 0), 1, "incorrect dotAt(1, 0)"},
-                {oBlock.dotAt(1, 1), 1, "incorrect dotAt(1, 1)"}
-        };
+    public void shouldBeRightQtOfColumns() {
+        assertEquals(block.cols(), image[0].length,
+                String.format("%s: Nr of columns for is incorrect.", block.getClass().getSimpleName()));
+    }
 
-        TBlock tBlock = new TBlock();
-        int actualRowsForTBlock = tBlock.rows();
-        int expectedRowsForTBlock = 2;
-        int actualColsForTBlock = tBlock.cols();
-        int expectedColsForTBlock = 3;
-        var TBlockDotRepresentation = new Object[][]{
-                {tBlock.dotAt(0, 0), 1, "incorrect dotAt(0, 0)"},
-                {tBlock.dotAt(0, 1), 1, "incorrect dotAt(0, 1)"},
-                {tBlock.dotAt(0, 2), 1, "incorrect dotAt(0, 2)"},
-                {tBlock.dotAt(1, 0), 0, "incorrect dotAt(1, 0)"},
-                {tBlock.dotAt(1, 1), 1, "incorrect dotAt(1, 1)"},
-                {tBlock.dotAt(1, 2), 0, "incorrect dotAt(1, 2)"},
-        };
-
-        YBlock yBlock = new YBlock();
-        int actualRowsForYBlock = yBlock.rows();
-        int expectedRowsForYBlock = 3;
-        int actualColsForYBlock = tBlock.cols();
-        int expectedColsForYBlock = 3;
-        var YBlockDotRepresentation = new Object[][]{
-                {yBlock.dotAt(0, 0), 1, "incorrect dotAt(0, 0)"},
-                {yBlock.dotAt(0, 1), 0, "incorrect dotAt(0, 1)"},
-                {yBlock.dotAt(0, 2), 1, "incorrect dotAt(0, 2)"},
-                {yBlock.dotAt(1, 0), 0, "incorrect dotAt(1, 0)"},
-                {yBlock.dotAt(1, 1), 1, "incorrect dotAt(1, 1)"},
-                {yBlock.dotAt(1, 2), 0, "incorrect dotAt(1, 2)"},
-                {yBlock.dotAt(2, 0), 0, "incorrect dotAt(2, 0)"},
-                {yBlock.dotAt(2, 1), 1, "incorrect dotAt(2, 1)"},
-                {yBlock.dotAt(2, 2), 0, "incorrect dotAt(2, 2)"}
-        };
-
-        JBlock jBlock = new JBlock();
-        int actualRowsForJBlock = jBlock.rows();
-        int expectedRowsForJBlock = 3;
-        int actualColsForJBlock = jBlock.cols();
-        int expectedColsForJBlock = 2;
-        var JBlockDotRepresentation = new Object[][]{
-                {jBlock.dotAt(0, 0), 0, "incorrect dotAt(0,0)"},
-                {jBlock.dotAt(0, 1), 1, "incorrect dotAt(0,1)"},
-                {jBlock.dotAt(1, 0), 0, "incorrect dotAt(1,0)"},
-                {jBlock.dotAt(1, 1), 1, "incorrect dotAt(1,1)"},
-                {jBlock.dotAt(2, 0), 1, "incorrect dotAt(2,0)"},
-                {jBlock.dotAt(2, 1), 1, "incorrect dotAt(2,1)"}
-        };
-
-        return new Object[][]{
-                {"Tests for OBlock " ,actualRowsForOBlock ,expectedRowsForOBlock ,actualColsForOBlock ,expectedColsForOBlock ,oBlockDotRepresentation},
-                {"Tests for TBlock " ,actualRowsForTBlock ,expectedRowsForTBlock ,actualColsForTBlock ,expectedColsForTBlock ,TBlockDotRepresentation},
-                {"Tests for YBlock " ,actualRowsForYBlock ,expectedRowsForYBlock ,actualColsForYBlock ,expectedColsForYBlock ,YBlockDotRepresentation},
-                {"Tests for JBlock " ,actualRowsForJBlock ,expectedRowsForJBlock ,actualColsForJBlock ,expectedColsForJBlock ,JBlockDotRepresentation}
-        };
+    @Test(dataProvider = "dotAtData", dependsOnMethods = {"shouldBeRightQtOfRows", "shouldBeRightQtOfColumns"})
+    public void shouldReturnRightBytesFromDotAtMethod(int i, int j, byte expectedDot) {
+        assertEquals(block.dotAt(i, j), expectedDot,
+                String.format("%s: incorrect dotAt(%s, %s)", block.getClass().getSimpleName(), i, j));
     }
 }
