+++ src/test/resources/testng.xml
@@ -11,7 +11,7 @@
             </run>
         </groups>
         <classes>
-            <class name="com.epam.prejap.tetris.block.BlockShapeTest"/>
+            <class name="com.epam.prejap.tetris.block.BlocksShapesTest"/>
         </classes>
     </test>
     <test verbose="2" name="jar tests">
